refactor: move signature utility functions from commands.core to utils
